+++ graphics/r_graphics.cpp
@@ -83,6 +83,7 @@ void Graphics::onResize(Vec2<> size) {
     //: Adjust render scale
     Vec2<float> ratios = win.size.to<float>() / Config::resolution.to<float>();
     win.scale = (ratios.x < ratios.y) ? floor(ratios.x) : floor(ratios.y);
+    setCameraProjection();
     
     //: Pass the resize command to the API
     API::resize(api, win);
@@ -133,16 +134,13 @@ TextureID Graphics::getTextureID(str path, Channels ch) {
 
     //: Load the image
     auto mode = [ch](){
-        switch(ch) {
-            case TEXTURE_CHANNELS_G:
-                return STBI_grey;
-            case TEXTURE_CHANNELS_GA:
-                return STBI_grey_alpha;
-            case TEXTURE_CHANNELS_RGB:
-                return STBI_rgb;
-            case TEXTURE_CHANNELS_RGBA:
-                return STBI_rgb_alpha;
-        }
+        if (ch == TEXTURE_CHANNELS_G)
+            return STBI_grey;
+        if (ch == TEXTURE_CHANNELS_GA)
+            return STBI_grey_alpha;
+        if (ch == TEXTURE_CHANNELS_RGB)
+            return STBI_rgb;
+        return STBI_rgb_alpha;
     }();
     
     Vec2<> size;
@@ -196,16 +194,23 @@ void Graphics::draw(const DrawID draw_id, glm::mat4 model) {
 }
 
 void Graphics::setCameraProjection() {
-    //: TODO: Improve and add options for perspective
-    camera.proj = glm::perspective(glm::radians(45.0f), win.size.x / (float) win.size.y, 0.1f, 10.0f);
+    //: Example perspective projection
+    //camera.proj = glm::perspective(glm::radians(45.0f), win.size.x / (float) win.size.y, 0.1f, 10.0f);
+    
+    //: Example orthographic projection (with (0,0) on the corner)
+    //camera.proj = glm::ortho(0.0f, win.size.x, 0.0f, win.size.y, -1000.0f, 1000.0f);
+    
+    //: Example orthographic projection (with (0,0) on the middle of the screen)
+    camera.proj = glm::ortho(-win.size.x/2.0f, win.size.x/2.0f, win.size.y/2.0f, -win.size.y/2.0f, -10000.0f, 10000.0f);
+    
     #ifdef USE_VULKAN //: Vulkan needs the Y direction flipped to match OpenGL
-    camera.proj[1][1] *= -1;
+    camera.proj[1][1] *= -1.0f;
     #endif
 }
 
 void Graphics::updateCameraView() {
-    //: Example of 2.5D camera, needs to be reworked for multiple camera view modes (2D, 3D...)
-    glm::vec3 camera_front = glm::vec3(0.0f, -0.2f, -1.0f);
-    glm::vec3 camera_pos = glm::vec3(camera.pos.x, camera.pos.y + 1.0f, 4.0f);
+    //: Example of view matrix
+    glm::vec3 camera_front = glm::vec3(0.0f, 0.0f, -1.0f);
+    glm::vec3 camera_pos = glm::vec3(camera.pos.x, camera.pos.y, 200.0f);
     camera.view = glm::lookAt(camera_pos, camera_pos + camera_front, glm::vec3(0.0f, 1.0f, 0.0f));
 }